Some new tiles that actually look nice!
I also added what I am calling a "Bouncy Growth." My idea is that it could be overlayed on tiles near the bouncy tiles to draw attention to them, no clue how it looks or if doable with the time left but the idea came to me nonetheless. Also made an example of this named Bouncy Growth Example.